test: delete assertions that cannot fail
Removes main.spec.ts (46 tests: 23 expect(true).toBe(true), 22 import-
existence checks already proven by the build and vue-tsc), preserving
the one real test in its own isolated file. Deletes a fake coverage-
summary describe block whose four tests were also expect(true).toBe(true).
Replaces three tolerant e2e assertions (count > 0 || count === 0,
try/catch swallowed into expect(true), an if-guarded assertion) with
strict either/or checks that can actually fail.

Also adds a missing setupBasicMocks(page) call to the plugin-auth e2e
test: without it the license-agreement modal blocked the page before
the assertion could even run.
