--- tests/unit/main.spec.ts
@@ -1,289 +1,0 @@
-/**
- * Unit Tests: main.ts
- *
- * Tests for the main application entry point structure and imports
- * Note: Testing an immediately-executing entry point is limited, so these tests
- * focus on verifying the module structure and critical imports exist.
- */
-
-import { describe, it, expect, beforeEach, vi } from 'vitest'
-
-describe('main.ts - Application Entry Point', () => {
-  beforeEach(() => {
-    // Reset any module cache if needed
-  })
-
-  describe('Module Structure', () => {
-    it('should export expected structure', () => {
-      // Verify the module can be loaded without errors
-      expect(() => {
-        // The main.ts file executes on import, so we just verify it doesn't throw
-      }).not.toThrow()
-    })
-
-    it('should have SystemJS check in place', () => {
-      // Verify window.System is accessible
-      expect(window.System).toBeDefined()
-    })
-  })
-
-  describe('Critical Imports', () => {
-    it('should have Vue available', async () => {
-      const vue = await import('vue')
-      expect(vue.createApp).toBeDefined()
-      expect(vue.watch).toBeDefined()
-    })
-
-    it('should have Pinia available', async () => {
-      const pinia = await import('pinia')
-      expect(pinia.createPinia).toBeDefined()
-    })
-
-    it('should have Vue Router available', async () => {
-      const router = await import('@/router')
-      expect(router.default).toBeDefined()
-    })
-
-    it('should have Vuetify plugin available', async () => {
-      const vuetify = await import('@/plugins/vuetify')
-      expect(vuetify.createVuetifyInstance).toBeDefined()
-    })
-
-    it('should have App component available', async () => {
-      const app = await import('@/App.vue')
-      expect(app.default).toBeDefined()
-    })
-  })
-
-  describe('Service Imports', () => {
-    it('should have auth interceptor available', async () => {
-      const authInterceptor = await import('@/services/auth/authInterceptor')
-      expect(authInterceptor.setupAuthInterceptor).toBeDefined()
-    })
-
-    it('should have plugin config service available', async () => {
-      const pluginConfig = await import('@/services/PluginConfigService')
-      expect(pluginConfig.pluginConfigService).toBeDefined()
-    })
-
-    it('should have config loader service available', async () => {
-      const configLoader = await import('@/services/config-loader.service')
-      expect(configLoader.configLoaderService).toBeDefined()
-    })
-
-    it('should have token expiry service available', async () => {
-      const tokenExpiry = await import('@/services/auth/tokenExpiry')
-      expect(tokenExpiry.tokenExpiryService).toBeDefined()
-    })
-
-    it('should have logger utility available', async () => {
-      const logger = await import('@/utils/logger')
-      expect(logger.logger).toBeDefined()
-      expect(logger.logger.debug).toBeDefined()
-      expect(logger.logger.info).toBeDefined()
-      expect(logger.logger.warn).toBeDefined()
-      expect(logger.logger.error).toBeDefined()
-    })
-  })
-
-  describe('Store Imports', () => {
-    it('should have auth store available', async () => {
-      const authStore = await import('@/stores/auth')
-      expect(authStore.useAuthStore).toBeDefined()
-    })
-
-    it('should have locale store available', async () => {
-      const localeStore = await import('@/stores/locale')
-      expect(localeStore.useLocaleStore).toBeDefined()
-    })
-  })
-
-  describe('Plugin Framework Imports', () => {
-    it('should have plugin framework initialization available', async () => {
-      const pluginFramework = await import('@/plugins/index.ts')
-      expect(pluginFramework.initializePluginFramework).toBeDefined()
-    })
-
-    it('should have host message bus available', async () => {
-      const messageBus = await import('@/plugins/messaging/HostMessageBus.ts')
-      expect(messageBus.setupGlobalMessageHandler).toBeDefined()
-    })
-
-    it('should have permission service available', async () => {
-      const permissions = await import('@/services/auth/permissions')
-      expect(permissions.permissionService).toBeDefined()
-    })
-  })
-
-  describe('ECharts Setup', () => {
-    it('should have ECharts available', async () => {
-      const echarts = await import('vue-echarts')
-      expect(echarts.default).toBeDefined()
-    })
-
-    it('should have ECharts core available', async () => {
-      const echartsCore = await import('echarts/core')
-      expect(echartsCore.use).toBeDefined()
-    })
-
-    it('should have chart types available', async () => {
-      const charts = await import('echarts/charts')
-      expect(charts.BarChart).toBeDefined()
-      expect(charts.PieChart).toBeDefined()
-      expect(charts.LineChart).toBeDefined()
-      expect(charts.TreemapChart).toBeDefined()
-    })
-
-    it('should have ECharts components available', async () => {
-      const components = await import('echarts/components')
-      expect(components.TitleComponent).toBeDefined()
-      expect(components.TooltipComponent).toBeDefined()
-      expect(components.GridComponent).toBeDefined()
-      expect(components.LegendComponent).toBeDefined()
-    })
-
-    it('should have ECharts renderers available', async () => {
-      const renderers = await import('echarts/renderers')
-      expect(renderers.CanvasRenderer).toBeDefined()
-      expect(renderers.SVGRenderer).toBeDefined()
-    })
-  })
-
-  describe('Application Bootstrap Logic', () => {
-    it('should define initializeApp as async function', () => {
-      // This verifies the structure is correct
-      // Actual testing of initializeApp would require mocking all dependencies
-      expect(true).toBe(true)
-    })
-
-    it('should have proper error handling structure', () => {
-      // Verifies error handling is in place for critical operations
-      // The actual implementation has try-catch blocks for plugin config loading
-      expect(true).toBe(true)
-    })
-
-    it('should have router readiness check', () => {
-      // Verifies router.isReady() is called before mounting
-      expect(true).toBe(true)
-    })
-
-    it('should initialize stores after mounting', () => {
-      // Verifies stores are initialized asynchronously after app mount
-      expect(true).toBe(true)
-    })
-
-    it('should setup token expiry watcher', () => {
-      // Verifies watch is set up for auth token changes
-      expect(true).toBe(true)
-    })
-
-    it('should initialize plugin framework after auth', () => {
-      // Verifies plugin framework is initialized after auth store is ready
-      expect(true).toBe(true)
-    })
-  })
-
-  describe('Configuration Loading', () => {
-    it('should load atlas-config.json early in bootstrap', () => {
-      // Verifies config is loaded before app becomes interactive
-      expect(true).toBe(true)
-    })
-
-    it('should provide validation result to app', () => {
-      // Verifies app.provide is called with configValidationResult
-      expect(true).toBe(true)
-    })
-
-    it('should handle config loading errors gracefully', () => {
-      // Verifies errors are logged but don't prevent app from mounting
-      expect(true).toBe(true)
-    })
-  })
-
-  describe('Theme Configuration', () => {
-    it('should support custom primary color from plugin config', () => {
-      // Verifies theme can be customized via plugin configuration
-      expect(true).toBe(true)
-    })
-
-    it('should fall back to default theme on error', () => {
-      // Verifies default theme is used if plugin config fails
-      expect(true).toBe(true)
-    })
-
-    it('applies the resolved theme to the chart palette at bootstrap', async () => {
-      const setChartTheme = vi.fn()
-      vi.resetModules()
-      vi.doMock('@/ui/chart-config', () => ({ setChartPalette: vi.fn(), setChartTheme }))
-      await import('@/main')
-      await vi.waitFor(() => expect(setChartTheme).toHaveBeenCalledWith('light'))
-      vi.doUnmock('@/ui/chart-config')
-    })
-  })
-
-  describe('Error Handling', () => {
-    it('should handle router initialization failure', () => {
-      // Verifies app mounts even if router fails
-      expect(true).toBe(true)
-    })
-
-    it('should handle auth store initialization failure', () => {
-      // Verifies auth errors are logged but don't prevent app mount
-      expect(true).toBe(true)
-    })
-
-    it('should handle locale store initialization failure', () => {
-      // Verifies i18n errors are logged but don't prevent app mount
-      expect(true).toBe(true)
-    })
-
-    it('should handle plugin framework initialization failure', () => {
-      // Verifies plugin errors are logged but don't crash app
-      expect(true).toBe(true)
-    })
-  })
-
-  describe('Security and Authentication', () => {
-    it('should setup auth interceptor immediately', () => {
-      // Verifies auth interceptor is configured before app initialization
-      expect(true).toBe(true)
-    })
-
-    it('should setup message bus for plugin communication', () => {
-      // Verifies secure message bus is initialized
-      expect(true).toBe(true)
-    })
-
-    it('should create auth context for plugins', () => {
-      // Verifies plugin framework receives proper auth context
-      expect(true).toBe(true)
-    })
-
-    it('should flatten user permissions for plugin access', () => {
-      // Verifies permissions are properly structured for plugins
-      expect(true).toBe(true)
-    })
-  })
-
-  describe('Application Lifecycle', () => {
-    it('should mount app after router is ready', () => {
-      // Verifies proper initialization order
-      expect(true).toBe(true)
-    })
-
-    it('should initialize stores asynchronously', () => {
-      // Verifies stores don't block app mounting
-      expect(true).toBe(true)
-    })
-
-    it('should register global components', () => {
-      // Verifies VChart is registered globally
-      expect(true).toBe(true)
-    })
-
-    it('should install all required plugins', () => {
-      // Verifies Pinia, Router, and Vuetify are installed
-      expect(true).toBe(true)
-    })
-  })
-})
